requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
attrs==18.2.0
Automat==0.7.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
Django==2.1.5
django-cors-headers==2.4.0
django-extensions==2.1.4
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.3.0
mccabe==0.6.1
parsel==1.5.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
pyOpenSSL==18.0.0
pytz==2018.9
queuelib==1.5.0
requests==2.21.0
Scrapy==1.5.1
scrapyd==1.2.0
scrapyd-client==1.1.0
service-identity==18.1.0
six==1.12.0
Twisted==18.9.0
urllib3==1.24.1
w3lib==1.20.0
wrapt==1.11.0
zope.interface==4.6.0
